Linter fixes
