+++ lib/nylas/api.rb
@@ -30,13 +30,13 @@ module Nylas
     # @return [CurrentAccount] The account details for whomevers access token is set
     def current_account
       prevent_calling_if_missing_access_token(:current_account)
-      CurrentAccount.from_hash(client.execute(method: :get, path: '/account'), api: self)
+      CurrentAccount.from_hash(client.execute(method: :get, path: "/account"), api: self)
     end
 
 
     private def prevent_calling_if_missing_access_token(method_name)
       return if client.access_token && !client.access_token.empty?
-      raise NoAuthToken, "No access token was provided and the #{method_name} method requires one"
+      raise NoAuthToken, method_name
     end
   end
 end
