+++ lib/nylas/http_client.rb
@@ -16,7 +16,9 @@ module Nylas
     # @return [Nylas::API]
     def initialize(app_id: , app_secret:, access_token: nil, api_server: 'https://api.nylas.com',
                    service_domain: 'api.nylas.com')
-      raise "When overriding the Nylas API server address, you must include https://" unless api_server.include?('://')
+      unless api_server.include?('://')
+        raise "When overriding the Nylas API server address, you must include https://"
+      end
       @api_server = api_server
       @access_token = access_token
       @app_secret = app_secret
@@ -47,27 +49,13 @@ module Nylas
       rest_client_execute(method: method, url: url, payload: payload,
                           headers: resulting_headers) do |response, request, result|
 
-        response = begin
-                     response.kind_of?(Enumerable) ? response : JSON.parse(response, symbolize_names: true)
-                   rescue JSON::ParserError
-                     response
-                   end
-        handle_failed_response(result: result, response: response, request: request)
+        response = parse_response(response)
+        handle_failed_response(result: result, response: response)
         response
       end
     end
     inform_on :execute, level: :debug,
       also_log: { result: true, values: [:method, :url, :path, :headers, :query, :payload] }
-
-    private def rest_client_execute(method: , url: , headers: , payload: , &block)
-      ::RestClient::Request.execute(method: method, url: url, payload: payload,
-                                    headers: headers, &block)
-    end
-    inform_on :rest_client_execute, level: :debug,
-      also_log: { result: true, values: [:method, :url, :headers, :payload] }
-
-
-
 
     # Syntactical sugar for making GET requests via the API.
     # @see #execute
@@ -90,7 +78,24 @@ module Nylas
     # Syntactical sugar for making DELETE requests via the API.
     # @see #execute
     def delete(path: nil, url: nil, payload: nil, headers: {}, query: {})
-      execute(method: :delete, path: path, url: url, headers: headers, query: query)
+      execute(method: :delete, path: path, url: url, headers: headers, query: query, payload: payload)
+    end
+
+
+
+    private def rest_client_execute(method: , url: , headers: , payload: , &block)
+      ::RestClient::Request.execute(method: method, url: url, payload: payload,
+                                    headers: headers, &block)
+    end
+    inform_on :rest_client_execute, level: :debug,
+      also_log: { result: true, values: [:method, :url, :headers, :payload] }
+
+    private def parse_response(response)
+      begin
+        response.kind_of?(Enumerable) ? response : JSON.parse(response, symbolize_names: true)
+      rescue JSON::ParserError
+        response
+      end
     end
 
     private def url_for_path(path)
@@ -99,17 +104,20 @@ module Nylas
       "#{protocol}//#{@access_token}:@#{domain}#{path}"
     end
 
-    private def handle_failed_response(result: , response:, request:)
+    private def handle_failed_response(result:, response:)
       http_code = result.code.to_i
 
+      handle_anticipated_failure_mode(http_code: http_code, response: response)
+      raise UnexpectedResponse.new(result.msg) if result.is_a?(Net::HTTPClientError)
+    end
+
+    private def handle_anticipated_failure_mode(http_code:, response:)
       if http_code != 200
-        HTTP_CODE_TO_EXCEPTIONS.fetch(http_code, APIError)
+        exception = HTTP_CODE_TO_EXCEPTIONS.fetch(http_code, APIError)
         if response.is_a?(Hash)
-          raise exc.new(response['type'], response['message'], response.fetch('server_error', nil))
+          raise exception.new(response['type'], response['message'], response.fetch('server_error', nil))
         end
       end
-
-      raise UnexpectedResponse.new(result.msg) if result.is_a?(Net::HTTPClientError)
     end
   end
 end
